delete dir alvaro and move class java to spaceinvader
