+++ src/main/java/org/fundacionjala/at15/spaceinvaders/Bullet.java
@@ -1,4 +1,4 @@
-package org.fundacionjala.at15.spaceinvaders.alvaro;
+package org.fundacionjala.at15.spaceinvaders;
 
 public class Bullet {
