+++ src/test/java/org/fundacionjala/at15/spaceinvaders/BulletTest.java
@@ -2,7 +2,7 @@ package org.fundacionjala.at15.spaceinvaders;
 
 import static org.junit.Assert.assertEquals;
 
-import org.fundacionjala.at15.spaceinvaders.alvaro.Bullet;
+import org.fundacionjala.at15.spaceinvaders.Bullet;
 import org.junit.Test;
 
 public class BulletTest {
@@ -14,7 +14,7 @@ public class BulletTest {
         Bullet bullet = new Bullet(posX, posY, fired);
         assertEquals(posX, bullet.getPosx());
         assertEquals(posY, bullet.getPosy());
-        
+
 
 
     }
